Create company
Students can now add a company directly through the Create Internship form. The submitted company information is first stored in the company_request table. After the guarantor approves the request, the company is moved to the main company table and becomes available in the company dropdown within the Create Internship form.